AO3-2452 More detailed printouts in tag tasks.

+++ app/models/common_tagging.rb
@@ -58,17 +58,17 @@ class CommonTagging < ApplicationRecord
   # Go through all CommonTaggings and destroy the invalid ones.
   def self.destroy_invalid
     includes(:common_tag, :filterable).find_each do |ct|
-      # Let callers do something on each iteration.
-      yield ct if block_given?
-
       # Don't use valid? because the uniqueness check triggers one query per row.
       # Just call the new checks, and check for the existence of errors (or
       # missing common_tag or filterable).
       ct.check_canonical_filterable
       ct.check_compatible_types
-      next if ct.errors.blank? && ct.common_tag && ct.filterable
+      valid = ct.errors.blank? && ct.common_tag && ct.filterable
 
-      ct.destroy
+      # Let callers do something on each iteration.
+      yield ct, valid if block_given?
+
+      ct.destroy unless valid
     end
   end
 end
